Conflicts merged and missing changes in new modules added as needed for clock co-scheduling.

+++ PDP18B/pdp18b_drm.c
@@ -244,7 +244,7 @@ static const int32 boot_rom[] = {
 
 t_stat drm_boot (int32 unitno, DEVICE *dptr)
 {
-int32 i;
+size_t i;
 extern int32 PC;
 
 if (drm_dib.dev != DEV_DRM)                             /* non-std addr? */
